CurrentSite: move site redirect notice to sub-component.
